MAINT: Install command-line tools via flit requires moving them to modules
